SIG-519 Moved user me view to it's own django app.
