+++ api/app/signals/apps/users/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from signals.apps.users.views import UserView
+
+urlpatterns = [
+    path('auth/me/', UserView.as_view()),
+]
