+++ api/app/signals/apps/users/views.py
@@ -1,0 +1,31 @@
+import re
+
+from django.http import JsonResponse
+from rest_framework.views import APIView
+
+from signals.apps.signals.views import AuthViewSet
+
+
+class UserView(AuthViewSet, APIView):
+    """
+    Handle information about user me
+    """
+    def get(self, request):
+        data = {}
+        user = request.user
+        if user:
+            data['username'] = user.username
+            data['email'] = user.email
+            data['is_staff'] = user.is_staff is True
+            data['is_superuser'] = user.is_superuser is True
+            groups = []
+            departments = []
+            for g in request.user.groups.all():
+                match = re.match(r"^dep_(\w+)$", g.name)
+                if match:
+                    departments.append(match.group(1))
+                else:
+                    groups.append(g.name)
+            data['groups'] = groups
+            data['departments'] = departments
+        return JsonResponse(data)
