+++ api/app/signals/urls.py
@@ -8,7 +8,10 @@ from django.urls import path
 
 urlpatterns = [
     path('status/', include('signals.apps.health.urls')),
+
+    # The Signals application is routed behind the HAproxy with `/signals/` as path.
     path('signals/', include('signals.apps.signals.urls')),
+    path('signals/auth/me', include('signals.apps.auth.urls')),
     path('signals/admin/', admin.site.urls),
 ]
